feat: 메인 페이지 서스펜스 적용해 streaming 동작하도록 변경
